Refactor frontend routing, fix API keys config, and resolve memory leaks.
- Renamed `src/ui/App.jsx` to `src/ui/Dashboard.jsx`.
- Created `src/ui/App.jsx` with `BrowserRouter` and `ProtectedRoute`.
- Updated API keys to use `REACT_APP_` prefix.
- Fixed memory leaks in `src/nlp/cd-detector.js` using `tf.tidy` and `dispose`.
- Added wildcard route in `server/index.js` for client-side routing.
- Enabled auth token interceptor in `src/services/api.js`.
- Fixed `webpack-dev-server` dependency conflict in `package.json`.
