+++ package.json
@@ -58,7 +58,6 @@
   },
   "overrides": {
     "nth-check": "^2.0.1",
-    "postcss": "^8.4.31",
-    "webpack-dev-server": "^5.2.0"
+    "postcss": "^8.4.31"
   }
 }
